refactoring Order Controller(separate api contents and impl
