+++ order/src/main/kotlin/kr/co/kimga/order/interfaces/exception/GlobalExceptionHandler.kt
@@ -1,10 +1,12 @@
 package kr.co.kimga.order.interfaces.exception
 
+import io.swagger.v3.oas.annotations.Hidden
 import org.springframework.http.HttpStatus
 import org.springframework.web.bind.annotation.ExceptionHandler
 import org.springframework.web.bind.annotation.ResponseStatus
 import org.springframework.web.bind.annotation.RestControllerAdvice
 
+@Hidden
 @RestControllerAdvice
 class GlobalExceptionHandler {
